neutral_sites: make ACGT-window filtering a one-time pinned step; evals_v2 consumes it (#270)
The N-window filter is model-independent (sites + genome + window), so re-running
it per checkpoint in evals_v2 was wasteful. Move it into the neutral_sites pipeline
next to pentanuc-annotation + subsampling, filter ONCE at the largest model window
(512) → one scoreable set reused by every model. evals_v2's calibration rule now
just reads the pre-filtered + subsampled set (no genome read, no per-model filter,
no fork-safety subprocess).

- sites.py: `filter_acgt_window_sites` (moved from evals/calibration.py; tests moved
  to test_sites.py). Drop sites whose centered window isn't all-ACGT (assembly-gap N
  or off-chromosome) — the scoring kernel asserts ACGT on the window; FWD checks the
  right flank, RC the left, so the whole window must be clean for both strands.
- neutral_sites rules: `stage_genome_fasta` (download the plain FASTA + .fai from S3
  — kept next to the .gz so genome-wide reads mmap a plain file instead of thrashing
  pyfaidx on bgzip), `filter_scoreable` (window `w` wildcard), and subsample now
  consumes the filtered set → `neutral_sites_n{n}_w{w}.parquet`. Filter BEFORE
  subsample so each 5-mer keeps a full n of clean sites.
- evals_v2 calibration: input is `neutral_sites_n{n}_w{scoreable_window}.parquet`;
  asserts scoreable_window >= the model's window_size. config: scoreable_window=512.
