+++ snakemake/neutral_sites/config/config.yaml
@@ -6,6 +6,13 @@
 # The S3 path requires `--group genome-s3` at install (pyfaidx + s3fs).
 genome_path: s3://oa-bolinas/data/genomes/homo_sapiens/GRCh38/ensembl-release-115/Homo_sapiens.GRCh38.dna_sm.primary_assembly.fa.gz
 
+# Plain (uncompressed) FASTA + .fai, kept in S3 next to the bgzipped reference.
+# The pentanuc / scoreable-window rules read the genome **genome-wide** (one whole-
+# chromosome span each), where pyfaidx on the bgzipped file is pathologically slow;
+# `stage_genome_fasta` downloads this plain file once and mmaps it instead. Produced
+# once with `zcat …fa.gz > …fa && python -c "import pyfaidx; pyfaidx.Faidx('…fa')"`.
+genome_fasta_path: s3://oa-bolinas/data/genomes/homo_sapiens/GRCh38/ensembl-release-115/Homo_sapiens.GRCh38.dna_sm.primary_assembly.fa
+
 # Source files — exactly what GPN-Star used for the hg38 calibration
 # (songlab-cal/gpn analysis/gpn-star/.../Snakefile_hg38 CALIBRATION_CONFIGS),
 # via UCSC https mirrors of GPN's ftp URLs (byte-identical files).
