+++ snakemake/neutral_sites/workflow/rules/subsample.smk
@@ -1,27 +1,50 @@
-"""Pentanucleotide annotation + per-5-mer subsampling for cLLR calibration (#267/#270).
+"""Pentanucleotide annotation, ACGT-window filtering, and per-5-mer subsampling
+for cLLR calibration (#267/#270).
 
-Both steps are model-independent (pure functions of genome + position + seed), so
-they live here rather than in the GPU eval pipeline. The subsampled artifact is
-pinned + seeded → every checkpoint calibrates on the same neutral sites, and it is
-kept site-level / alt-agnostic so the LLR path (3 alts) and the entropy atom
-(#269, 4-allele marginal) consume one set.
+All three steps are model-independent (pure functions of genome + position +
+window + seed), so the calibration-ready set is built and pinned **here**, once,
+and reused by every checkpoint in the GPU eval pipeline (evals_v2) — never
+re-filtered per model. The artifact is seeded + kept site-level / alt-agnostic so
+the LLR path (3 alts) and the entropy atom (#269, 4-allele marginal) consume one set.
+
+Order matters: filter to scoreable sites **before** subsampling, so each 5-mer's
+``n`` sites are drawn from its *clean* sites (a full ``n`` per cell, not ``n`` minus
+whatever the filter later removes).
+
+Genome reads go through a plain (uncompressed) FASTA (`stage_genome_fasta`):
+pyfaidx random access on the bgzipped reference is pathologically slow for the
+genome-wide spans these rules touch (one whole-chromosome read each), so we mmap a
+plain file instead. The plain FASTA + its `.fai` are kept in S3 next to the
+bgzipped reference (`genome_fasta_path`) so this is a one-time *download*, never a
+re-decompress.
 """
+
+
+rule stage_genome_fasta:
+    """Download the plain (uncompressed) reference FASTA + .fai from S3 for fast,
+    genome-wide pyfaidx access (mmap). The uncompressed reference lives in S3 next
+    to the bgzipped one precisely so this is a download, not a per-run `zcat`.
+    Local-only scratch (`temp`): the ~3 GB file is removed once its consumers finish."""
+    output:
+        temp("results/genome.fa"),
+    shell:
+        "aws s3 cp {config[genome_fasta_path]} {output} && "
+        "aws s3 cp {config[genome_fasta_path]}.fai {output}.fai"
 
 
 rule annotate_pentanuc:
     """Attach each neutral site's central pentanucleotide (5-mer) from the genome.
 
-Additive (consumes the pinned neutral set, leaves it untouched); the genome is
-already wired into this pipeline, so context extraction belongs here."""
+Additive (consumes the pinned neutral set, leaves it untouched); reads the plain
+FASTA once per chromosome and asserts the 5-mer center == ref."""
     input:
-        "results/neutral_sites.parquet",
+        sites="results/neutral_sites.parquet",
+        genome="results/genome.fa",
     output:
         "results/neutral_sites_pentanuc.parquet",
     run:
-        sites = pd.read_parquet(input[0])
-        # Genome is the get_seq callable (same as enumerate_positions); reads
-        # once per chromosome and asserts the 5-mer center == ref.
-        out = annotate_pentanucleotide(sites, Genome(config["genome_path"]))
+        sites = pd.read_parquet(input.sites)
+        out = annotate_pentanucleotide(sites, Genome(input.genome))
         assert len(out) > 0, "empty pentanuc-annotated neutral set"
         out.to_parquet(output[0], index=False)
         print(
@@ -30,17 +53,45 @@ already wired into this pipeline, so context extraction belongs here."""
         )
 
 
-rule subsample_neutral:
-    """Subsample to at most {n} neutral sites per 5-mer (seeded, deterministic).
+rule filter_scoreable:
+    """Drop sites whose centered {w}-bp model window isn't all-ACGT (assembly-gap N
+    or off-chromosome edge).
 
-`n` is a path wildcard so several caps coexist (e.g. n100, n1000) and evals_v2
-pins one by revision — see README."""
+The gLM scoring kernel asserts ACGT over the variant's window, so an N trips it;
+filtering here (once) lets every checkpoint reuse one scoreable set. `w` is a path
+wildcard — use the largest model window (512) so the set is valid for all models (a
+window clean at 512 bp is clean at 255/256). Runs before `subsample_neutral`."""
     input:
-        "results/neutral_sites_pentanuc.parquet",
+        sites="results/neutral_sites_pentanuc.parquet",
+        genome="results/genome.fa",
     output:
-        "results/subsampled/neutral_sites_n{n}.parquet",
+        "results/scoreable/neutral_sites_pentanuc_w{w}.parquet",
+    wildcard_constraints:
+        w=r"\d+",
+    run:
+        sites = pd.read_parquet(input.sites)
+        out = filter_acgt_window_sites(sites, Genome(input.genome), int(wildcards.w))
+        assert len(out) > 0, "every site filtered out — wrong window/genome?"
+        out.to_parquet(output[0], index=False)
+        print(
+            f"[filter_scoreable] w={wildcards.w}: {len(out):,} scoreable sites "
+            f"({out['pentanuc'].nunique()} distinct 5-mers) -> {output[0]}"
+        )
+
+
+rule subsample_neutral:
+    """Subsample to at most {n} scoreable sites per 5-mer (seeded, deterministic),
+    from the window-{w} ACGT-filtered set.
+
+`n` and `w` are path wildcards so several caps / windows coexist; evals_v2 pins one
+(by `subsample_n` + `scoreable_window`) — see README."""
+    input:
+        "results/scoreable/neutral_sites_pentanuc_w{w}.parquet",
+    output:
+        "results/subsampled/neutral_sites_n{n}_w{w}.parquet",
     wildcard_constraints:
         n=r"\d+",
+        w=r"\d+",
     run:
         sites = pd.read_parquet(input[0])
         n_cap = int(wildcards.n)
@@ -49,8 +100,8 @@ pins one by revision — see README."""
         n_depleted = int((per < n_cap).sum())  # 5-mers with fewer than n sites
         out.to_parquet(output[0], index=False)
         print(
-            f"[subsample_neutral] n={n_cap} seed={config['subsample_seed']}: "
-            f"{len(out):,} sites across {len(per)} 5-mers "
-            f"(<= {per.max()} per 5-mer; {n_depleted} 5-mers below the cap, "
+            f"[subsample_neutral] n={n_cap} w={wildcards.w} "
+            f"seed={config['subsample_seed']}: {len(out):,} sites across {len(per)} "
+            f"5-mers (<= {per.max()} per 5-mer; {n_depleted} 5-mers below the cap, "
             f"min {per.min()}) -> {output[0]}"
         )
