+++ tests/pipelines/neutral_sites/test_sites.py
@@ -13,6 +13,7 @@ from marin_dna.pipelines.neutral_sites.sites import (
     annotate_pentanucleotide,
     contiguous_runs,
     enumerate_positions,
+    filter_acgt_window_sites,
     neutral_mask,
     parse_rmsk,
     scan_neutral_intervals,
@@ -335,3 +336,55 @@ class TestSubsamplePerContext:
         src = self._sites()
         out = subsample_per_context(src, n=100, seed=0)
         assert len(out) == len(src)
+
+
+class _FakeGenome:
+    """Minimal ``get_seq(chrom, start, end)`` with N-padding past the end (like Genome).
+
+    ``filter_acgt_window_sites`` clamps the read start to >= 0, so ``start`` is never
+    negative here.
+    """
+
+    def __init__(self, seqs: dict[str, str]):
+        self.seqs = seqs
+
+    def __call__(self, chrom: str, start: int, end: int) -> str:
+        seg = self.seqs[chrom][start:end]
+        return seg + "N" * ((end - start) - len(seg))  # right-pad N past chrom end
+
+
+class TestFilterAcgtWindowSites:
+    def test_drops_n_and_edge_windows(self):
+        # 30 bp chrom with N at 0-based indices 10 and 25; window_size=5 → left_flank=2,
+        # so site pos's window is 0-based [pos-3, pos+2).
+        seq = list("ACGTACGTACGTACGTACGTACGTACGTAC")
+        seq[10] = "N"
+        seq[25] = "N"
+        seq = "".join(seq)
+        genome = _FakeGenome({"1": seq})
+        sites = pd.DataFrame(
+            {
+                "chrom": ["1"] * 6,
+                # 5: clean | 8: clean | 13: N upstream (FWD-ok/RC-fail case) |
+                # 24: N downstream | 2: window off left edge | 29: window off right edge
+                "pos": [5, 8, 13, 24, 2, 29],
+                "ref": [seq[p - 1] for p in [5, 8, 13, 24, 2, 29]],
+            }
+        )
+        out = filter_acgt_window_sites(sites, genome, window_size=5)
+        assert set(out["pos"]) == {5, 8}
+        assert list(out.columns) == list(sites.columns)  # columns/order preserved
+
+    def test_all_clean_kept(self):
+        genome = _FakeGenome({"1": "ACGT" * 50})  # 200 bp, no N
+        sites = pd.DataFrame({"chrom": ["1"] * 3, "pos": [50, 100, 150]})
+        out = filter_acgt_window_sites(sites, genome, window_size=11)
+        assert len(out) == 3
+
+    def test_requires_columns(self):
+        with pytest.raises(AssertionError, match="pos"):
+            filter_acgt_window_sites(
+                pd.DataFrame({"chrom": ["1"]}),
+                _FakeGenome({"1": "ACGT"}),
+                window_size=5,
+            )
